RaspberryPi/requirements.txt
numpy==1.24.2
opencv_contrib_python==4.7.0.72
Pillow==9.5.0
pyserial==3.5
multiprocess==0.70.14
pipreqs==0.4.11
